Move recently played logic from page to component
